requirements.txt
streamlit==1.51.0
pandas==2.3.3
pillow==12.0.0
rembg==2.0.68
qrcode==8.2
PyMuPDF==1.26.6
openpyxl==3.1.5